feat(mobile): 导航重构(导入进档案/导出分享独立 tab)+ 设置瘦身
按用户产品反馈重排(手机端定位「轻」:采集+看+分享;搜索/趋势归桌面与查看器):
- 底部导航 [健康档案 / 导出分享 / 设置],默认停「健康档案」。
- 导入并进「健康档案」右上角「+ 导入」→ 弹三选一(拍照/相册/选文件),导入后经
  vaultRevision 自动刷新;逻辑抽到 import_flow.dart(进度改模态对话框)。
- 导出/分享独立成 export_screen.dart(时间线导出带日期筛选 + 加密分享给医生)。
- 删旧 import_export_screen.dart;Patrol 测试对齐新导航(app_test + export_test)。
- 设置瘦身:清空行去掉与确认弹窗重复的灰字(_SettingsRow.subtitle 改可空)。
版本 1.2.0+5。
