+++ apps/mobile_flutter/lib/main.dart
@@ -3,8 +3,8 @@ import 'package:path_provider/path_provider.dart';
 import 'package:mobile_flutter/src/rust/frb_generated.dart';
 import 'package:mobile_flutter/src/rust/api/vault.dart';
 import 'package:mobile_flutter/theme.dart';
-import 'package:mobile_flutter/screens/import_export_screen.dart';
 import 'package:mobile_flutter/screens/archive_screen.dart';
+import 'package:mobile_flutter/screens/export_screen.dart';
 import 'package:mobile_flutter/screens/settings_screen.dart';
 
 Future<void> main() async {
@@ -105,11 +105,9 @@ class HomeShell extends StatefulWidget {
 class _HomeShellState extends State<HomeShell> {
   int _index = 0;
 
-  static const _screens = [
-    ImportExportScreen(),
-    ArchiveScreen(),
-    SettingsScreen(),
-  ];
+  // 健康档案(看 + 右上角导入)· 导出分享 · 设置。导入并进「健康档案」,
+  // 导出/分享独立成 tab —— 手机端「轻」定位:采集 + 看 + 分享,搜索/趋势在桌面/查看器。
+  static const _screens = [ArchiveScreen(), ExportScreen(), SettingsScreen()];
 
   @override
   Widget build(BuildContext context) {
@@ -120,14 +118,14 @@ class _HomeShellState extends State<HomeShell> {
         onDestinationSelected: (i) => setState(() => _index = i),
         destinations: const [
           NavigationDestination(
-            icon: Icon(Icons.swap_vert_outlined),
-            selectedIcon: Icon(Icons.swap_vert),
-            label: '导入导出',
-          ),
-          NavigationDestination(
             icon: Icon(Icons.folder_outlined),
             selectedIcon: Icon(Icons.folder),
             label: '健康档案',
+          ),
+          NavigationDestination(
+            icon: Icon(Icons.ios_share_outlined),
+            selectedIcon: Icon(Icons.ios_share),
+            label: '导出分享',
           ),
           NavigationDestination(
             icon: Icon(Icons.settings_outlined),
